Clone procedure object within each Writer to avoid modification
Previously ProcedureWriter.procedure was a reference to the procedure
object. However, if one ProcedureWriter modified that procedure object,
then other ProcedureWriters may be affected. Instead now each
ProcedureWriter uses lodash\cloneDeep to ensure it has its own copy of
the procedure object.

This was recognized to be an issue when `pat build --sodf` was run and
the SODF procedure was a valid Word document but the EVA procedure could
not be opened. Running just `pat build` to build the EVA procedure
produced a valid Word document.