Run HTML screenshot test against Complex Times case; check image size
Comparing the Complex Times to the "simple" screenshot only saw a ~3.5%
difference, despite being very different files. Comparing Complex Times
against a random _very_ different image only was ~15% difference. So the
percentages aren't totally to be trusted. It'd be better to run these
tests in a container to get a consistent result.