Add ability to render PNGs from SVGs in browser (specifically timeline)
Two packages have to be removed by Webpack to run SVG conversion in
browser: svg2img and svgdom. Both of these have conflicts with the
browser (node native canvas versus browser canvas and mockup of a window
object versus an actual browser window object). In the case of svgdom,
just remove the unneeded module (replace with empty object). In the case
of svg2img, use a browser-friendly function instead.

While this work was predominantly done to make it possible to render
summary timelines from SVGs into PNGs in the browser, this will likely
have future usefulness in importing images in the editor UI, and perhaps
annotating them.